# contracts/openapi/notification-gateway.openapi.yaml
openapi: 3.0.3
info:
  title: Notification Gateway Query API
  description: >-
    Read-only query API over the notification record. Exposes a single-record
    lookup by id and a paged search by delivery status and creation date.
    Interim posture: served open (unauthenticated) until authentication is
    added; the query API is read-only (no command/write side effects).
  version: 1.0.0
servers:
  - url: /
paths:
  /notifications/{notificationId}:
    get:
      operationId: getNotificationById
      summary: Retrieve a single notification record by its id
      parameters:
        - name: notificationId
          in: path
          required: true
          description: The notification identifier.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The notification record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationView'
        '400':
          description: >-
            The request is invalid — e.g. a malformed notification id. No response body is returned.
        '404':
          description: No notification exists for the supplied id.
  /notifications:
    get:
      operationId: searchNotifications
      summary: Search notification records by status and creation date
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by delivery status.
          schema:
            $ref: '#/components/schemas/NotificationStatus'
        - name: createdFrom
          in: query
          required: false
          description: Inclusive lower bound of the notification creation timestamp.
          schema:
            type: string
            format: date-time
        - name: createdTo
          in: query
          required: false
          description: Inclusive upper bound of the notification creation timestamp.
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          required: false
          description: Zero-based page index.
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          description: Page size.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 200
            default: 20
      responses:
        '200':
          description: A page of matching notification records.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPage'
        '400':
          description: >-
            The request has invalid parameters — e.g. an out-of-range page or size, an unknown status,
            or a malformed date-time. No response body is returned.
components:
  schemas:
    NotificationStatus:
      type: string
      description: The lifecycle status of a notification.
      enum:
        - QUEUED
        - SENT
        - FAILED
    NotificationView:
      type: object
      description: A notification record — all columns of the notification table.
      required:
        - notificationId
        - notificationType
        - status
        - createdAt
        - updatedAt
      properties:
        notificationId:
          type: string
          format: uuid
        notificationType:
          type: string
        status:
          type: string
        sendToAddress:
          type: string
          nullable: true
        statusCode:
          type: integer
          format: int32
          nullable: true
        errorMessage:
          type: string
          nullable: true
        clientContext:
          type: string
          nullable: true
        resultQueue:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NotificationPage:
      type: object
      description: A page of notification records.
      required:
        - content
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/NotificationView'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32